finishing reorganization
